Fixing period markers not spanning enough rows.
At least part of the problem was that the
when the firstNRows were removed it only
moved the year marker down if the month
also needed moving down. But months don't
need moving down at the end of the month
so this meant the year got lost from the
start of the table.
When new rows were then added back they
didn't find the year label so created
one but with a rowspan of 1, where it should
have been more. Hence the missing
rows.

This fixes the year, but not the month issue.